CaptionBar
Restored the button on the caption bar.
The dashboard now updates when the data server is started / stopped.
Calculations on F9 have been optimed to reduce the load when running with many different entry points
The data refresh has been optimised.
Fixed an issue when Sophis loads and the FindRootLoadedPortfolio function